backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
#mysqlclient==2.2.0
PyMySQL==1.1.2

